Add a script to test effect of a noisey oil profile.
It is likely that the oil mass ratio depth profile from a coupled model
would be noisey and on a coarse grid. So simulate this and interpolate
to pass to infinite layer model to see if it is an issue for the solver.

Upshot is the irradaince profiles remain smooth but the radiative
heating output becomes noisey reflecting the noisey input absorption
which should be ok.