calcuate absorption and extinction coeffs using interpolated roche data
We remove the previous implementation using data from an otremba paper
and the associated script demonstrating it.

The oil MAC data for Romashkino oil provided by Roche is interpolated in
wavelength and droplet radius space and used to calculate the ice
absoprtion and extinction coefficients.

Note that we must add the extra parameter to specify the median droplet
radius of a log normally distribution of oil droplets with standard
geometric deviation of e.