Make median_droplet_radius_in_microns a parameter for models.
Where the models are run in the scripts to plot some diagnostics we need
to also provide a median droplet radius for the oil absorption
calculation

Also allow for the calculation of absorption outside of roche oil data
range for wavelengths >800nm. Beyond these wavelengths ice absorption
increases and so the effect of oil is less important so just interpolate
the MAC as zero here.