change interface of inifinite layer model to maximise efficiency.
The parameters for solution are now specified in the InfiniteLayerModel
object which now takes the dimensional discretised grid to evaluate the
solution on as well as the discretised points in wavelength space.

The model is solved and returns a SpectralIrradiance object which
provides irradiances in a 2D array of vertical coordinate and
wavelength.

I've also written a function to integrate this object in wavelength
space and return an Irradiance object with the total integrated
irradiances.