Implement black body spectrum as a Callable class.
This makes the most sense as an implementation as we can now easily get
a normalised black body spectrum in a given SW wavelength range.

Change the integrate_over_SW function to require the spectrum over which
to integrate seperately. This means it is decoupled from the wavelength
range in the spectral irradiance object so we can integrate over the
desired spectrum independently.