Add fast_solve option to infinite layer model.
This requires specifying a wavelength cutoff. For short wavelengths run
the ode solver. But for longer wavelengths approximate the solution
as all absorbed within the first grid point. This cutoff therefore
depends on the grid resolution the solution is requested on and the
extinction coefficient in the domain.

For resolutons of around 1cm a cutoff of 1500nm seems appropriate